requirements.txt
# Equipment Database System Dependencies

# PostgreSQL Database Drivers
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Database ORM and Utilities
SQLAlchemy==2.0.23
alembic==1.13.1

# Web Framework
Flask==3.0.0

# Async Support
asyncio-mqtt==0.16.1

# Logging and Configuration
python-dotenv==1.0.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code Quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0
